+++ src/main/java/org/gudy/azureus2/pluginsimpl/local/messaging/GenericMessageConnectionIndirect.java
@@ -393,7 +393,7 @@ GenericMessageConnectionIndirect
 
                 } else {
 
-                    i = i.intValue() + 1;
+                    i = i + 1;
                 }
 
                 totals.put(originator, i);
